Fix live version of WaniKani.
The lessons are loaded in batches of several items and when
switching page new data is updated via javascript and not by
loading a new web page. Set up a listener that listens to
changes to the main-info element and calls init() whenever
it changes.

Also need to have different ids for the show/hide links for
radical/kanji/vocabulary as the HTML for all three types are
avaialbe at the same time during lessons.